refactor: modularize editor islands
