+++ src/islands/definitions-tree/constants.ts
@@ -1,0 +1,7 @@
+export const listSelectors = {
+    node: '.definition-node',
+    item: '.definition-item',
+    actions: '.definition-actions',
+};
+
+export const definitionsTargetSelector = '#definitions-list';
